refactored webpack, jest resource configs; also adds more tasks to just
