+++ apps/server-rendered-app/webpack.serve.config.js
@@ -1,5 +1,5 @@
 const path = require('path');
-const resources = require('../../scripts/tasks/webpack-resources');
+const resources = require('../../scripts/webpack/webpack-resources');
 const isProduction = process.argv.indexOf('--production') > -1;
 const PACKAGE_NAME = 'todo-app';
