+++ apps/ssr-tests/webpack.config.js
@@ -1,5 +1,5 @@
 const path = require('path');
-const resources = require('../../scripts/tasks/webpack-resources');
+const resources = require('../../scripts/webpack/webpack-resources');
 
 module.exports = resources.createConfig('ssr-tests', false, {
   entry: './test/test.js',
@@ -18,14 +18,8 @@ module.exports = resources.createConfig('ssr-tests', false, {
     alias: {
       'office-ui-fabric-react/src': path.join(__dirname, 'node_modules/office-ui-fabric-react/src'),
       'office-ui-fabric-react/lib': path.join(__dirname, 'node_modules/office-ui-fabric-react/lib'),
-      '@uifabric/fabric-website-resources/src': path.join(
-        __dirname,
-        'node_modules/@uifabric/fabric-website-resources/src'
-      ),
-      '@uifabric/fabric-website-resources/lib': path.join(
-        __dirname,
-        'node_modules/@uifabric/fabric-website-resources/lib'
-      ),
+      '@uifabric/fabric-website-resources/src': path.join(__dirname, 'node_modules/@uifabric/fabric-website-resources/src'),
+      '@uifabric/fabric-website-resources/lib': path.join(__dirname, 'node_modules/@uifabric/fabric-website-resources/lib'),
       '@uifabric/styling/lib': path.join(__dirname, '../../packages/styling/lib'),
       'Props.ts.js': 'Props'
     }
