+++ package.json
@@ -33,9 +33,9 @@
     "checkchange": "node common/scripts/install-run-rush.js change -v",
     "update-api": "cd packages/office-ui-fabric-react && npm run update-api",
     "update-snapshots": "cd packages/office-ui-fabric-react && npm run update-snapshots",
-    "check-for-changed-files": "node ./scripts/check-for-modified-files.js",
+    "check-for-changed-files": "node scripts/just.js check-for-modified-files",
     "prettier": "node scripts/prettier.js",
-    "generate-version-files": "node scripts/generate-version-files.js",
+    "generate-version-files": "node scripts/just.js generate-version-files",
     "codepen": "cd packages/office-ui-fabric-react && node ../../scripts/local-codepen.js",
     "precommit": "node scripts/node_modules/lint-staged/index.js"
   },
