+++ packages/foundation-scenarios/webpack.serve.config.js
@@ -1,5 +1,5 @@
 const path = require('path');
-const resources = require('../../scripts/tasks/webpack-resources');
+const resources = require('../../scripts/webpack/webpack-resources');
 const webpack = resources.webpack;
 
 const PACKAGE_NAME = require('./package.json').name;
