+++ packages/merge-styles/webpack.config.js
@@ -1,5 +1,5 @@
 const path = require('path');
-const resources = require('../../scripts/tasks/webpack-resources');
+const resources = require('../../scripts/webpack/webpack-resources');
 const webpack = resources.webpack;
 
 const BUNDLE_NAME = 'merge-styles';
