+++ packages/office-ui-fabric-react/webpack.codepen.config.js
@@ -1,5 +1,5 @@
 const path = require('path');
-const resources = require('../../scripts/tasks/webpack-resources');
+const resources = require('../../scripts/webpack/webpack-resources');
 
 module.exports = resources.createServeConfig({
   entry: './src/index.bundle.ts',
@@ -18,10 +18,7 @@ module.exports = resources.createServeConfig({
   resolve: {
     alias: {
       'office-ui-fabric-react/src': path.resolve(__dirname, '../../packages/office-ui-fabric-react/src'),
-      'office-ui-fabric-react/lib/codepen': path.resolve(
-        __dirname,
-        '../../packages/office-ui-fabric-react/lib/codepen'
-      ),
+      'office-ui-fabric-react/lib/codepen': path.resolve(__dirname, '../../packages/office-ui-fabric-react/lib/codepen'),
       'office-ui-fabric-react/lib': path.resolve(__dirname, '../../packages/office-ui-fabric-react/src'),
       'Props.ts.js': 'Props',
       'Example.tsx.js': 'Example'
