+++ packages/theme-samples/jest.config.js
@@ -1,3 +1,3 @@
-let { createConfig } = require('../../scripts/tasks/jest-resources');
+let { createConfig } = require('../../scripts/jest/jest-resources');
 
 module.exports = createConfig();
