+++ packages/webpack-utils/jest.config.js
@@ -1,5 +1,5 @@
 const path = require('path');
-const { createConfig } = require('../../scripts/tasks/jest-resources');
+const { createConfig } = require('../../scripts/jest/jest-resources');
 
 module.exports = createConfig({
   setupFiles: [path.resolve(path.join(__dirname, 'config', 'tests.js'))]
