+++ scripts/templates/create-package/EmptyWebpackConfig.mustache
@@ -1,5 +1,5 @@
 const path = require('path');
-const resources = require('../../scripts/tasks/webpack-resources');
+const resources = require('../../scripts/webpack/webpack-resources');
 
 const BUNDLE_NAME = '{{{packageName}}}';
 const IS_PRODUCTION = process.argv.indexOf('--production') > -1;
